Add `--last-failed` option (#895)
Adding the `--last-failed` option to protostar so it may work similarly
as in `pytest`.

By the way, I added new util that allows managing protostar's cache.
It's very simple as I needed it just for this one case but it is
obviously open to improvements if necessary.

So now, if someone uses this flag, only the tests that previously failed
will be launched. If on the last run all tests passed, again all of them
will be launched. The data about which tests passed is always collected
(through the cache) but used only if the `--last-failed` flag is
present.

#693