Fixed bug for builtins and memory holes (#1192)
It fixes the bug that profiler blamed instruction address for a samples
instead of callstack. The same instruction can be executed from
different callstacks e.g library function. As a result all sample were
assigned to the one, last callstack of the instruction.
Instead such tree
```
a -> b -> d
| -> c ----|
```
Execution resulted in
```
a -> c -> d(2)
```