Change scope of errors reraised in bindings (#1669)
I discovered that there are errors that are thrown from rust and are not
`RuntimeError`. The result was, in my case, an infinite hang which is
very bad. The error was something like `pyo3_runtime.PanicException`.
Are there any objections to just catching everything that comes from
rust (I mean errors) and turning it into `CairoBindingException`?