[CI] Further cost optimization (#1953)
This one is for further cost optimizations on CI. 
It turned out mac checks ate significant chunk of our credits, so from
now on we will run all the checks only on master to make sure release is
stable. For PRs there will be only linux checks. + Added pre-caching for
mac jobs, to further optimize credit-time for jobs that do not yet have
cache created.

### Checklist
- [X] Relevant issue is linked
- [X] Docs updated/issue for docs created
- [X] Added relevant tests